added error handling

+++ server/index.ts
@@ -1,17 +1,17 @@
-import express from 'express';
+/// <reference path="./types/custom.d.ts" />
+import express, { Request, Response, NextFunction } from 'express';
 import cors from 'cors';
 import pool from './db';
 import bcrypt from 'bcrypt';
-import jwt from 'jsonwebtoken';
+import jwt, { JwtPayload } from 'jsonwebtoken';
 
 const app = express();
 
-//middleware
 app.use(cors());
 app.use(express.json());
 
-const JWT_SECRET = '123aBcDe4FDS'
-
+const JWT_SECRET = '123aBcDe4FDS';
+  
 type Booking = {
     eid: number;
     title: string;
@@ -21,24 +21,65 @@ type Booking = {
     resource?: boolean; 
 };
 
-type UpdateBooking = {
-    eid: number;
-    title?: string;  
-    start?: string;  
-    end_time?: string;
-    allDay?: boolean;
-    resource?: boolean;
+interface UpdateBooking extends Partial<Booking> {}
+
+interface ExtendedJwtPayload extends JwtPayload {
+    userId: number;
+    email: string;
+  }
+  
+const authenticateToken = (req: Request, res: Response, next: NextFunction) => {
+    const authHeader = req.headers.authorization;
+    const token = authHeader?.split(' ')[1];
+  
+    if (!token) {
+      return res.sendStatus(401);
+    }
+  
+    jwt.verify(token, JWT_SECRET, (err, user) => {
+        if (err) {
+            console.log("JWT Verification Error:", err);
+            return res.sendStatus(403);
+        }
+        req.user = user as ExtendedJwtPayload;
+        next();
+    });
 };
 
 //ROUTES//
 
 // API POST endpoint to create a booking
-app.post("/bookings", async (req, res) => {
+app.post("/bookings", authenticateToken, async (req, res) => {
     try {
-        const { eid, title, start, end_time, allDay, resource }: Booking = req.body;
-        const insertValues: Array<string | number | boolean> = [eid, title, start, end_time];
-        const valuePlaceholders = ["$1", "$2", "$3", "$4"];
-        const insertColumns = ["eid", "title", "start", "end_time"];
+        const userId = req.user?.userId; 
+        if (!userId) {
+           return res.status(403).json({ message: "User ID missing from request" });
+        }
+
+        const bookingStartDate = new Date(req.body.start);
+        const firstDayOfBookingMonth = new Date(bookingStartDate.getFullYear(), bookingStartDate.getMonth(), 1);
+        const lastDayOfBookingMonth = new Date(bookingStartDate.getFullYear(), bookingStartDate.getMonth() + 1, 0);
+
+        const countQuery = `
+            SELECT COUNT(*) 
+            FROM bookings 
+            WHERE resident_id = $1 AND start >= $2 AND end_time <= $3
+        `;
+        const countResult = await pool.query(countQuery, [userId, firstDayOfBookingMonth.toISOString(), lastDayOfBookingMonth.toISOString()]);
+        const bookingsCount = parseInt(countResult.rows[0].count);
+
+        if (bookingsCount >= 3) {
+            return res.status(400).json({ message: "You have already made 3 bookings for the booking month." });
+        }
+
+        const maxEidResult = await pool.query('SELECT MAX(eid) FROM bookings');
+        const maxEid = maxEidResult.rows[0].max || 0;
+        const newEid = maxEid + 1;
+
+        const { title, start, end_time, allDay, resource } = req.body;
+        const insertValues = [newEid, title, start, end_time, userId];
+        const valuePlaceholders = ["$1", "$2", "$3", "$4", "$5"];
+        const insertColumns = ["eid", "title", "start", "end_time", "resident_id"];
 
         if (allDay !== undefined) {
             insertColumns.push("allDay");
@@ -52,12 +93,12 @@ app.post("/bookings", async (req, res) => {
         }
 
         const query = `INSERT INTO bookings (${insertColumns.join(", ")}) VALUES (${valuePlaceholders.join(", ")}) RETURNING *`;
-
         const newBooking = await pool.query(query, insertValues);
 
         res.json(newBooking.rows[0]);
     } catch (err) {
         console.error(err);
+        res.status(500).json({ message: "Server error" });
     }
 });
 
@@ -71,7 +112,7 @@ app.get('/bookings', async (req, res) => {
     }
 });
 
-// API GET endpoint to get a single booking
+// API GET endpoint to get a single booking using booking id (eid)
 app.get('/bookings/:eid', async (req, res) => {
     try {
         const { eid } = req.params;
@@ -86,46 +127,81 @@ app.get('/bookings/:eid', async (req, res) => {
 });
 
 // API PUT endpoint to update a single booking
-app.put("/bookings/:eid", async (req, res) => {
+app.put("/bookings/:eid", authenticateToken, async (req: Request, res: Response) => {
+    const { eid } = req.params;
+    const { title, start, end_time, allDay, resource }: UpdateBooking = req.body;
+    const userId = req.user?.userId;
+
+    if (!userId) {
+        return res.status(403).json({ message: "User ID missing from request" });
+    }
+    if (!eid || !start || !end_time) {
+        return res.status(400).json("Missing required fields");
+    }
+
+    const bookingResult = await pool.query('SELECT * FROM bookings WHERE eid = $1', [eid]);
+    if (bookingResult.rows.length === 0) {
+        return res.status(404).json({ message: "Booking not found" });
+    }
+    const booking = bookingResult.rows[0];
+    if (booking.resident_id !== userId) {
+        return res.status(403).json({ message: "Unauthorized to update this booking" });
+    }
+
+    const newStartDate = new Date(start);
+    const firstDayOfMonth = new Date(newStartDate.getFullYear(), newStartDate.getMonth(), 1);
+    const lastDayOfMonth = new Date(newStartDate.getFullYear(), newStartDate.getMonth() + 1, 0);
+
+    const countQuery = `
+        SELECT COUNT(*) 
+        FROM bookings 
+        WHERE resident_id = $1 AND start >= $2 AND end_time <= $3 AND eid != $4
+    `;
+    const countResult = await pool.query(countQuery, [userId, firstDayOfMonth, lastDayOfMonth, eid]);
+    const bookingsCount = parseInt(countResult.rows[0].count);
+
+    if (bookingsCount >= 3) {
+        return res.status(400).json({ message: "Updating this booking will exceed your limit of 3 bookings for the month." });
+    }
+
     try {
-        const { eid } = req.params;
-        const { title, start, end_time, allDay, resource }: UpdateBooking = req.body;
-
-        if (!eid || !start || !end_time) {
-            return res.status(400).json("Missing required fields");
-        }
-
+        const updateValues: any[] = [];
         const updateFields: string[] = [];
-        const updateValues: (string | number | boolean)[] = [];
 
         if (title !== undefined) {
             updateFields.push("title = $1");
             updateValues.push(title);
         }
-
-        updateFields.push("start = $2", "end_time = $3");
-        updateValues.push(start, end_time);
-
-        if (allDay != null) {
-            updateFields.push("allDay = $4");
+        if (start !== undefined) {
+            updateFields.push("start = $" + (updateValues.length + 1));
+            updateValues.push(start);
+        }
+        if (end_time !== undefined) {
+            updateFields.push("end_time = $" + (updateValues.length + 1));
+            updateValues.push(end_time);
+        }
+        if (allDay !== undefined) {
+            updateFields.push("allday = $" + (updateValues.length + 1));
             updateValues.push(allDay);
         }
-        if (resource != null) {
-            updateFields.push("resource = $5");
+        if (resource !== undefined) {
+            updateFields.push("resource = $" + (updateValues.length + 1));
             updateValues.push(resource);
         }
 
-        updateValues.push(eid);
+        console.log("Update values:", updateValues);
+        updateValues.push(eid, userId);
 
-        const updateQuery = `UPDATE bookings SET ${updateFields.join(", ")} WHERE eid = $${updateValues.length}`;
-        console.log(updateQuery);
+        const updateQuery = `
+            UPDATE bookings 
+            SET ${updateFields.join(", ")} 
+            WHERE eid = $${updateValues.length - 1} AND resident_id = $${updateValues.length}
+        `;
 
         await pool.query(updateQuery, updateValues);
 
         const updatedBooking = await pool.query("SELECT * FROM bookings WHERE eid = $1", [eid]);
-
         res.json(updatedBooking.rows[0]);
-        console.log(updatedBooking.rows[0]);
     } catch (err) {
         console.error(err);
         res.status(500).send('Server error');
@@ -133,44 +209,89 @@ app.put("/bookings/:eid", async (req, res) => {
 });
 
 // API DELETE endpoint to remove a booking
-app.delete("/bookings/:eid", async (req, res) => {
+app.delete("/bookings/:eid", authenticateToken, async (req: Request, res: Response) => {
+    const { eid } = req.params;
+    const userId = parseInt(req.user?.userId);
+
+    const bookingResult = await pool.query('SELECT * FROM bookings WHERE eid = $1', [parseInt(eid)]);
+
+    if (bookingResult.rows.length === 0) {
+        return res.status(404).json({ message: "Booking not found" });
+    }
+    const booking = bookingResult.rows[0];
+
+    if (booking.resident_id !== userId) {
+        return res.status(403).json({ message: "Unauthorized to delete this booking" });
+    }
+
     try {
-        const { eid } = req.params;
-        await pool.query("DELETE FROM bookings WHERE eid = $1", [eid]);
-        res.json("booking was deleted!");
+        await pool.query("DELETE FROM bookings WHERE eid = $1 AND resident_id = $2", [parseInt(eid), userId]);
+        res.json({ message: "Booking deleted successfully" });
     } catch (err) {
-        console.log(err);
+        res.status(500).json({ message: 'Server error' });
     }
 });
 
 // API POST endpoint for residents to login
-app.post('/login', async (req, res) => {
+app.post('/login', async (req: Request, res: Response) => {
     const { email, password } = req.body;
   
     try {
       const userResult = await pool.query('SELECT * FROM residents WHERE email = $1', [email]);
-
       if (userResult.rows.length === 0) {
         return res.status(401).json({ message: 'Authentication failed' });
       }
   
       const user = userResult.rows[0];
-  
       const isMatch = await bcrypt.compare(password, user.password);
-
       if (!isMatch) {
         return res.status(401).json({ message: 'Authentication failed' });
       }
   
-      const token = jwt.sign({ userId: user.resident_id, email: user.email }, JWT_SECRET, { expiresIn: '1h' });
-
+      const token = jwt.sign({ userId: user.resident_id, email: user.email }, JWT_SECRET, { expiresIn: '3h' });
       res.json({ token });
     } catch (error) {
-      console.error('Error during login:', error);
+      console.error(error);
       res.status(500).json({ message: 'Server error' });
     }
 });
 
+// API GET endpoint to get resident information using resident_id
+app.get('/residents', authenticateToken, async (req, res) => {
+    try {
+        const userId = req.user?.userId;
+
+        if (!userId) {
+            return res.status(403).json({ message: "User ID missing from request" });
+        }
+
+        const userInfo = await pool.query('SELECT * FROM residents WHERE resident_id = $1', [userId]);
+        res.json(userInfo.rows);
+    } catch (err) {
+        console.log(err);
+    }
+});
+
+// API GET endpoint to get all bookings for a single resident using resident_id
+app.get('/userbookings', authenticateToken, async (req, res) => {
+    try {
+        const userId = req.user?.userId;
+
+        if (!userId) {
+            return res.status(403).json({ message: "User ID missing from request" });
+        }
+
+        const bookings = await pool.query('SELECT * FROM bookings WHERE resident_id = $1', [userId]);
+        res.json(bookings.rows);
+    } catch (err) {
+        console.log(err);
+    }
+});
+
+// API POST endpoint for residents to logout
+app.post('/logout', authenticateToken, (req: Request, res: Response) => {
+    res.status(200).json({ message: 'Logout successful' });
+});
 
 app.listen(5000, () => {
     console.log("server has started on port 5000");
